ingestion/phase-5-retrieval/requirements.txt
chromadb>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
sentence-transformers>=3.0.0,<4.0.0
numpy>=1.26.0,<3.0.0
